Move statement parsing to a separate file
